requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
autoflake==2.3.1
black==24.4.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
com2ann==0.3.0
discord.py==2.3.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
iso8601==2.1.0
libcst==1.4.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PlexAPI==4.15.13
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pyflakes==3.2.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyupgrade==3.16.0
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
rsa==4.9
ruff==0.4.10
shed==2024.3.1
six==1.16.0
spotipy==2.24.0
tokenize-rt==5.2.0
tomli==2.0.1
twitchAPI==4.2.0
twitchio==2.9.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
yt-dlp==2024.5.27
